+++ src/app/app.module.ts
@@ -1,13 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { RouterModule } from '@angular/router';
-import { PageLoaderComponent, ProductComponent } from './components';
 
 @NgModule({
-  imports: [BrowserModule, CommonModule, RouterModule.forRoot([])],
-  declarations: [AppComponent, PageLoaderComponent, ProductComponent],
+  imports: [BrowserModule, CommonModule, AppRoutingModule],
+  declarations: [AppComponent],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
