+++ src/app/products/components/index.ts
