+++ src/app/products/components/page-loader/page-loader.component.ts
