+++ src/app/products/components/product/product.component.scss
