+++ src/app/products/models/index.ts
