+++ src/app/products/products.component.scss
@@ -1,0 +1,47 @@
+.products {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(16rem, 1fr));
+  gap: 20px;
+  padding: 20px;
+}
+
+.product-link {
+  display: flex;
+
+  &:link,
+  &:visited,
+  &:hover,
+  &:active {
+    color: unset;
+    text-decoration: unset;
+  }
+}
+
+.more {
+  font: inherit;
+  display: block;
+  margin: 4rem auto 0;
+  border: 1px solid #d7dbe0;
+  border-radius: 0.25rem;
+  padding: 1rem 2rem;
+  background-color: #fff;
+  cursor: pointer;
+
+  &:hover,
+  &:focus {
+    background-color: #ebecf0;
+  }
+}
+
+.error-box {
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  z-index: 999;
+  background-color: red;
+  color: white;
+  padding: 1rem;
+  font-size: 1rem;
+  font-weight: bold;
+}
