+++ src/app/products/states/index.ts
@@ -1,0 +1,1 @@
+export * from './products.state';
